+++ giford/image/__init__.py
@@ -1,3 +1,3 @@
-from .abstract_frame_wrapper import AbstractFrameWrapper
+from .abstract_image import AbstractImage
 from .single_image import SingleImage, SingleImageFormat
 from .multi_image import MultiImage, MultiImageFormat
