+++ giford/image/abstract_image.py
@@ -4,7 +4,7 @@ from giford.frame_batch import FrameBatch
 from giford.raw_data import RawDataFrame
 
 
-class AbstractFrameWrapper(abc.ABC):
+class AbstractImage(abc.ABC):
     """
     container for raw data frames
     also acts as interface from dataframe <-> file on disk
