+++ giford/image/multi_image.py
@@ -9,7 +9,7 @@ from PIL import Image as PillowImage
 
 from giford.frame_batch import FrameBatch
 from giford.raw_data import RawDataFrame, RawDataVideo
-from giford.frame_wrapper.abstract_frame_wrapper import AbstractFrameWrapper
+from giford.image.abstract_image import AbstractImage
 
 
 class MultiImageFormat(enum.Enum):
@@ -22,7 +22,7 @@ class MultiImageFormat(enum.Enum):
     GIF = 1
 
 
-class MultiImage(AbstractFrameWrapper):
+class MultiImage(AbstractImage):
     DEFAULT_FRAMERATE = 15
     DEFAULT_FORMAT = MultiImageFormat.GIF
