+++ tests/test_shake.py
@@ -4,7 +4,7 @@ import pytest
 
 from giford.frame_batch import FrameBatch
 from giford.image_actions.shake import Shake
-from giford.frame_wrapper import MultiImage
+from giford.image import MultiImage
 from tests.util import BASELINE_DIRECTORY, save_batch_and_compare
