+++ tests/test_single_image.py
@@ -1,4 +1,4 @@
-from giford.frame_wrapper import AbstractFrameWrapper, SingleImage
+from giford.image import AbstractImage, SingleImage
 from giford.frame_batch import FrameBatch
 from tests.util import compare_image_files, TEST_INPUT_ORANGE_IMAGE_FILEPATH
