+++ tests/util.py
@@ -5,7 +5,7 @@ import numpy as np
 from PIL import Image as PillowImage
 
 from giford.frame_batch import FrameBatch
-from giford.frame_wrapper import (
+from giford.image import (
     SingleImage,
     SingleImageFormat,
     MultiImage,
